Resolving failing unit tests, added coverage for OwnerMappingProfile

+++ 4. Accessors/MusicAPI.Accessors/DataTransferObjects/BaseSearchResult.cs
@@ -14,7 +14,7 @@ namespace MusicAPI.Accessors.DataTransferObjects
         /// The maximum number of items in the reponse (as set in the query or by default).
         /// </summary>
         [JsonPropertyName("limit")]
-        public int Limit { get; set; } = 0;
+        public decimal Limit { get; set; } = 0;
 
         /// <summary>
         /// URL to the next page of items. (null if none)
@@ -26,7 +26,7 @@ namespace MusicAPI.Accessors.DataTransferObjects
         /// The offset of the items returned (as set in the query or by default).
         /// </summary>
         [JsonPropertyName("offset")]
-        public int Offset { get; set; } = 0;
+        public decimal Offset { get; set; } = 0;
 
         /// <summary>
         /// URL to the previous page of items (null if none).
@@ -38,6 +38,6 @@ namespace MusicAPI.Accessors.DataTransferObjects
         /// The total number of items available to return.
         /// </summary>
         [JsonPropertyName("total")]
-        public int Total { get; set; } = 0;
+        public decimal Total { get; set; } = 0;
     }
 }
